requirements.txt
psutil==5.9.8
pyserial==3.5